🩹 change checkin_size to positive small Int

+++ reception/read_qr/models.py
@@ -8,7 +8,7 @@ class Reservation(models.Model):
     name = models.CharField(max_length=100)
     name_kana = models.CharField(max_length=100)
     mail = models.CharField(max_length=200)
-    party_size = models.IntegerField(default=0)
+    party_size = models.PositiveSmallIntegerField(default=0)
     reserved_at = models.DateTimeField("datetime reserved")
     notes = models.CharField(blank=True, max_length=500)
 
@@ -25,7 +25,7 @@ class Reservation(models.Model):
 
 class Checkin(models.Model):
     reservation = models.ForeignKey(Reservation, on_delete=models.CASCADE)
-    checkin_size = models.IntegerField(default=0)
+    checkin_size = models.PositiveSmallIntegerField(default=0)
     checkin_at = models.DateTimeField("datetime checkin-ed")
     notes = models.CharField(blank=True, max_length=500)
